Fix #73: "Install" after opening the Solution did nothing
