+++ src/ExtensionManager.Shared/Commands/SolutionPrompter.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+
 using Microsoft.VisualStudio.Shell;
 
 namespace ExtensionManager
@@ -8,9 +9,11 @@ namespace ExtensionManager
     public class SolutionPrompter
     {
         private readonly ExtensionService _extService;
+        private readonly Package _package;
 
-        public SolutionPrompter(ExtensionService extService)
+        public SolutionPrompter(Package package, ExtensionService extService)
         {
+            _package = package;
             _extService = extService;
         }
 
@@ -30,7 +33,13 @@ namespace ExtensionManager
             if (manifest.Extensions.Any(e => e.Selected))
             {
                 var msg = "This solution asks that you install the following extensions.";
-                Importer.ImportWindow.Open(manifest.Extensions, Importer.Purpose.InstallForSolution, msg);
+                var dialog = Importer.ImportWindow.Open(manifest.Extensions, Importer.Purpose.InstallForSolution, msg);
+
+                if (dialog.DialogResult == true && dialog.SelectedExtension.Any())
+                {
+                    new ExtensionInstaller(_package)
+                        .Install(dialog.SelectedExtension, dialog.InstallSystemWide);
+                }
             }
         }
     }
