+++ src/ExtensionManager.Shared/ExtensionManager.Shared.projitems
@@ -13,6 +13,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Commands\ExportSolutionCommand.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Commands\ImportCommand.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Commands\SolutionPrompter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ExtensionInstaller.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExtensionService.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Importer\ImportWindow.xaml.cs">
       <DependentUpon>ImportWindow.xaml</DependentUpon>
